% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{Send a postfix email}
\usage{
send_email(to, subject, body)
}
\arguments{
\item{to}{email address of the recipient}

\item{subject}{subject of email}

\item{body}{body of the email}
}
\description{
Send a postfix email
}
