% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_long_email.R
\name{send_long_email}
\alias{send_long_email}
\title{Send a long postfix email in a vector of lines exported to a tmp text file}
\usage{
send_long_email(to, subject, vector)
}
\arguments{
\item{to}{email address of the recipient}

\item{subject}{subject of email}

\item{vector}{vector to compose body of the email}
}
\description{
Send a long postfix email in a vector of lines exported to a tmp text file
}
